requirements.txt
opencv-python~=4.12.0.88
torch~=2.9.1
albumentations~=2.0.8
torchvision~=0.24.1
matplotlib~=3.10.7
tqdm~=4.67.1
numpy~=2.2.6
scikit-learn~=1.7.2
streamlit~=1.51.0
pillow~=12.0.0